Moved commands into utils
